refactor package with figures
