+++ chessboard/src/main/java/com/github/fertilewaif/chessboard/model/pieces/Queen.kt
@@ -1,4 +1,7 @@
-package com.github.fertilewaif.chessboard.model
+package com.github.fertilewaif.chessboard.model.pieces
+
+import com.github.fertilewaif.chessboard.model.Board
+import com.github.fertilewaif.chessboard.model.CellInfo
 
 class Queen(isWhite: Boolean) : Piece(isWhite) {
     val rook = Rook(isWhite)
